Update OBJ1 -Nikon40x0.95air-.txt
Slowed down PID to remove oscillations and increased the acceptable error value for the FOCUS flag from 0.05 (default) to 0.1. The FOCUS flag now seems stable 'ON' when settled in a new XY stage location after some fast motion. Sample flag value reduced to sensible level so that the flag is 'ON' when a sample is present and 'OFF' when the sample is removed. These settings may only be valid for the current objective and sample combination...