Update prior_PureFocus850.py
Improved error checking and more informative error messages on 'set_piezo_voltage' with the aim of catching when the method is called and the autofocus 'servo' is enabled (should not be done). This can happen either from a programmatic error (method calls in wrong order) or because the user has pressed a button on the controller. An approach here could be to get/set the servo every time when calling 'set_piezo_voltage', but this incurs a ~10ms time penalty (so not good for z-stacks). So instead we try to highlight the error so it can be corrected (i.e. program correctly, or more defensively, or leave the buttons alone). To help with this the state of the servo is now checked during the init.